Separate native functions and add forward action
