+++ modules/ballerina-http/src/main/java/org/ballerinalang/net/http/HTTPServicesRegistry.java
@@ -244,7 +244,7 @@ public class HTTPServicesRegistry {
         //Validate request parameter
         if (isValidParam(paramDetails.get(1), Constants.REQUEST)) {
             throw new BallerinaConnectorException("second parameter should be of type - "
-                    + Constants.PROTOCOL_PACKAGE_HTTP + ":" + Constants.REQUEST);
+                    + Constants.PROTOCOL_PACKAGE_HTTP + ":" + Constants.IN_REQUEST);
         }
 
         //validate rest of the parameters
