+++ modules/ballerina-http/src/main/java/org/ballerinalang/net/http/nativeimpl/connection/Respond.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.net.http.nativeimpl.outbound.response;
+package org.ballerinalang.net.http.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.connector.api.ConnectorUtils;
